refactor: move BulkSelectionStore to global, it is not bankTransactions-only
Putting it under bankTransactions was wrong, and not just as a matter of taste:
blocks/BulkActionsModule consumes it, and a domain-agnostic block may not
depend on a feature domain. The earlier placement pointed a block at
@providers/bankTransactions and inverted the layering.

Neither file has anything to do with bank transactions. The store is keyed on
bare ids (useSelectedIds, useIdIsSelected), and useMobileListBulkSelection
takes an id array and adapts the store to react-aria's Selection API. Both are
generic, so global/ is where they belong regardless of who happens to consume
them today.

Note that no lint rule caught this. The layer ordering is enforced for
@ui/@blocks/@features imports but not for a block reaching into a domain's
providers, so the violation was silent.
