+++ src/components/features/bankTransactions/BankTransactions/BankTransactions.tsx
@@ -13,10 +13,10 @@ import { BankTransactionsFiltersProvider } from '@providers/bankTransactions/Ban
 import { BankTransactionsPaginationProvider } from '@providers/bankTransactions/BankTransactionsPagination/BankTransactionsPaginationProvider'
 import { BankTransactionsRouteStoreProvider } from '@providers/bankTransactions/BankTransactionsRouteStore/BankTransactionsRouteStoreProvider'
 import { BankTransactionsStringOverridesProvider } from '@providers/bankTransactions/BankTransactionsStringOverridesContext/BankTransactionsStringOverridesContext'
-import { BulkSelectionStoreProvider } from '@providers/bankTransactions/BulkSelectionStore/BulkSelectionStoreProvider'
 import { BankTransactionsCategorizationStoreProvider } from '@providers/categorization/BankTransactionsCategorizationStore/BankTransactionsCategorizationStoreProvider'
 import { BankTransactionsIsCategorizationEnabledProvider } from '@providers/categorization/BankTransactionsIsCategorizationEnabledContext/BankTransactionsIsCategorizationEnabledContext'
 import { CategorizationRulesProvider } from '@providers/categorization/CategorizationRulesContext/CategorizationRulesContext'
+import { BulkSelectionStoreProvider } from '@providers/global/BulkSelectionStore/BulkSelectionStoreProvider'
 import { InAppLinkProvider, type LinkingMetadata } from '@providers/global/InAppLink/InAppLinkContext'
 import { type BankTransactionsMode, LegacyModeProvider } from '@providers/global/LegacyModeProvider/LegacyModeProvider'
 import { ErrorBoundary } from '@components/utility/ErrorBoundary/ErrorBoundary'
