+++ src/components/features/bankTransactions/BankTransactionsBulkActions/BankTransactionsCategorizeAllModal.tsx
@@ -9,8 +9,8 @@ import { useBulkCategorizeBankTransactions } from '@hooks/features/bankTransacti
 import { useTaxCodeOptions } from '@hooks/features/bankTransactions/useTaxCodeOptions'
 import { useIntlFormatter } from '@hooks/utils/i18n/useIntlFormatter'
 import { useBankTransactionsContext } from '@providers/bankTransactions/BankTransactions/BankTransactionsContext'
-import { useBulkSelectionActions, useCountSelectedIds, useSelectedIds } from '@providers/bankTransactions/BulkSelectionStore/BulkSelectionStoreProvider'
 import type { BankTransactionNonSuggestedMatchOption } from '@providers/categorization/BankTransactionsCategorizationStore/utils'
+import { useBulkSelectionActions, useCountSelectedIds, useSelectedIds } from '@providers/global/BulkSelectionStore/BulkSelectionStoreProvider'
 import { VStack } from '@ui/Stack/Stack'
 import { Label, Span } from '@ui/Typography/Text'
 import { BaseConfirmationModal } from '@blocks/BaseConfirmationModal/BaseConfirmationModal'
