+++ src/components/features/bankTransactions/BankTransactionsBulkActions/BankTransactionsConfirmAllModal.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from 'react-i18next'
 import { tPlural } from '@utils/i18n/plural'
 import { useBulkMatchOrCategorize } from '@hooks/features/bankTransactions/useBulkMatchOrCategorize'
 import { useIntlFormatter } from '@hooks/utils/i18n/useIntlFormatter'
-import { useBulkSelectionActions, useCountSelectedIds } from '@providers/bankTransactions/BulkSelectionStore/BulkSelectionStoreProvider'
+import { useBulkSelectionActions, useCountSelectedIds } from '@providers/global/BulkSelectionStore/BulkSelectionStoreProvider'
 import { VStack } from '@ui/Stack/Stack'
 import { Span } from '@ui/Typography/Text'
 import { BaseConfirmationModal } from '@blocks/BaseConfirmationModal/BaseConfirmationModal'
