+++ src/components/features/bankTransactions/BankTransactionsBulkActions/BankTransactionsUncategorizeAllModal.tsx
@@ -4,8 +4,8 @@ import { useTranslation } from 'react-i18next'
 import { tPlural } from '@utils/i18n/plural'
 import { useBulkUncategorizeBankTransactions } from '@hooks/features/bankTransactions/useBulkBankTransactionMutations'
 import { useIntlFormatter } from '@hooks/utils/i18n/useIntlFormatter'
-import { useBulkSelectionActions, useCountSelectedIds, useSelectedIds } from '@providers/bankTransactions/BulkSelectionStore/BulkSelectionStoreProvider'
 import { useBankTransactionsCategorizationActions } from '@providers/categorization/BankTransactionsCategorizationStore/BankTransactionsCategorizationStoreProvider'
+import { useBulkSelectionActions, useCountSelectedIds, useSelectedIds } from '@providers/global/BulkSelectionStore/BulkSelectionStoreProvider'
 import { Span } from '@ui/Typography/Text'
 import { BaseConfirmationModal } from '@blocks/BaseConfirmationModal/BaseConfirmationModal'
