+++ src/components/features/bankTransactions/BankTransactionsHeader/BankTransactionsHeader.tsx
@@ -13,8 +13,8 @@ import { useBankTransactionsContext } from '@providers/bankTransactions/BankTran
 import { BankTransactionsFeature, useIsBankTransactionsFeatureEnabled } from '@providers/bankTransactions/BankTransactionsFeatureVisibility/BankTransactionsFeatureVisibilityProvider'
 import { useBankTransactionsFiltersContext } from '@providers/bankTransactions/BankTransactionsFiltersContext/BankTransactionsFiltersContext'
 import { useBankTransactionsStringOverrides } from '@providers/bankTransactions/BankTransactionsStringOverridesContext/BankTransactionsStringOverridesContext'
-import { useCountSelectedIds } from '@providers/bankTransactions/BulkSelectionStore/BulkSelectionStoreProvider'
 import { useBankTransactionsIsCategorizationEnabledContext } from '@providers/categorization/BankTransactionsIsCategorizationEnabledContext/BankTransactionsIsCategorizationEnabledContext'
+import { useCountSelectedIds } from '@providers/global/BulkSelectionStore/BulkSelectionStoreProvider'
 import { useGlobalDateRangeActions } from '@providers/global/DateStoreProvider/GlobalDateStoreProvider'
 import { MonthPicker } from '@ui/DatePickers/MonthPicker/MonthPicker'
 import { HStack, VStack } from '@ui/Stack/Stack'
