+++ src/components/features/bankTransactions/BankTransactionsList/BankTransactionsListItem.tsx
@@ -17,9 +17,9 @@ import { useSaveBankTransactionRow } from '@hooks/features/bankTransactions/useS
 import { useIntlFormatter } from '@hooks/utils/i18n/useIntlFormatter'
 import { useSizeClass } from '@hooks/utils/size/useWindowSize'
 import { useBankTransactionsStringOverrides } from '@providers/bankTransactions/BankTransactionsStringOverridesContext/BankTransactionsStringOverridesContext'
-import { useBulkSelectionActions, useIdIsSelected } from '@providers/bankTransactions/BulkSelectionStore/BulkSelectionStoreProvider'
 import { useBankTransactionsCategorizationActions } from '@providers/categorization/BankTransactionsCategorizationStore/BankTransactionsCategorizationStoreProvider'
 import { useBankTransactionsIsCategorizationEnabledContext } from '@providers/categorization/BankTransactionsIsCategorizationEnabledContext/BankTransactionsIsCategorizationEnabledContext'
+import { useBulkSelectionActions, useIdIsSelected } from '@providers/global/BulkSelectionStore/BulkSelectionStoreProvider'
 import { SubmitAction } from '@ui/Button/SubmitButton'
 import { Checkbox } from '@ui/Checkbox/Checkbox'
 import { Chevron } from '@ui/Chevron/Chevron'
