+++ src/components/features/bankTransactions/BankTransactionsMobileList/BankTransactionsMobileList.tsx
@@ -5,8 +5,8 @@ import { type BankTransaction } from '@internal-types/bankTransactions'
 import { useUpsertBankTransactionsDefaultCategories } from '@hooks/features/bankTransactions/useUpsertBankTransactionsDefaultCategories'
 import { useBankTransactionsContext } from '@providers/bankTransactions/BankTransactions/BankTransactionsContext'
 import { useBankTransactionsFiltersContext } from '@providers/bankTransactions/BankTransactionsFiltersContext/BankTransactionsFiltersContext'
-import { useBulkSelectionActions } from '@providers/bankTransactions/BulkSelectionStore/BulkSelectionStoreProvider'
-import { useMobileListBulkSelection } from '@providers/bankTransactions/BulkSelectionStore/useMobileListBulkSelection'
+import { useBulkSelectionActions } from '@providers/global/BulkSelectionStore/BulkSelectionStoreProvider'
+import { useMobileListBulkSelection } from '@providers/global/BulkSelectionStore/useMobileListBulkSelection'
 import { VStack } from '@ui/Stack/Stack'
 import { MobileList } from '@blocks/MobileList/MobileList'
 import { useMobileListExpansion } from '@blocks/MobileList/useMobileListExpansion'
