+++ src/components/features/bankTransactions/BankTransactionsTable/BankTransactionCategoryCell.tsx
@@ -10,9 +10,9 @@ import { useDelayedRemoveBankTransaction } from '@hooks/features/bankTransaction
 import { useGetBankTransactionMatchOrCategoryWithDefault } from '@hooks/features/bankTransactions/useGetBankTransactionCategorizationWithDefault'
 import { useSaveBankTransactionRow } from '@hooks/features/bankTransactions/useSaveBankTransactionRow'
 import { useBankTransactionsStringOverrides } from '@providers/bankTransactions/BankTransactionsStringOverridesContext/BankTransactionsStringOverridesContext'
-import { useBulkSelectionActions, useCountSelectedIds } from '@providers/bankTransactions/BulkSelectionStore/BulkSelectionStoreProvider'
 import { useBankTransactionsCategorizationActions } from '@providers/categorization/BankTransactionsCategorizationStore/BankTransactionsCategorizationStoreProvider'
 import { useBankTransactionsIsCategorizationEnabledContext } from '@providers/categorization/BankTransactionsIsCategorizationEnabledContext/BankTransactionsIsCategorizationEnabledContext'
+import { useBulkSelectionActions, useCountSelectedIds } from '@providers/global/BulkSelectionStore/BulkSelectionStoreProvider'
 import { Button } from '@ui/Button/Button'
 import { SubmitAction } from '@ui/Button/SubmitButton'
 import { Chevron } from '@ui/Chevron/Chevron'
