+++ src/components/features/bankTransactions/BankTransactionsTable/BankTransactionsTable.tsx
@@ -13,8 +13,8 @@ import { useBankTransactionsContext } from '@providers/bankTransactions/BankTran
 import { useBankTransactionsFiltersContext } from '@providers/bankTransactions/BankTransactionsFiltersContext/BankTransactionsFiltersContext'
 import { useBankTransactionsPaginationContext } from '@providers/bankTransactions/BankTransactionsPagination/BankTransactionsPaginationContext'
 import { useBankTransactionsStringOverrides } from '@providers/bankTransactions/BankTransactionsStringOverridesContext/BankTransactionsStringOverridesContext'
-import { useBulkSelectionActions, useSelectedIds } from '@providers/bankTransactions/BulkSelectionStore/BulkSelectionStoreProvider'
 import { useBankTransactionsIsCategorizationEnabledContext } from '@providers/categorization/BankTransactionsIsCategorizationEnabledContext/BankTransactionsIsCategorizationEnabledContext'
+import { useBulkSelectionActions, useSelectedIds } from '@providers/global/BulkSelectionStore/BulkSelectionStoreProvider'
 import { MoneySpan } from '@ui/Typography/MoneySpan'
 import { Span } from '@ui/Typography/Text'
 import type { ClickableRowProps } from '@blocks/DataTable/DataTable'
