+++ src/hooks/features/bankTransactions/useBulkMatchOrCategorize.ts
@@ -7,8 +7,8 @@ import {
   type MatchOrCategorizeTransactionRequestSchema,
 } from '@schemas/bankTransactions/bulkMatchOrCategorize'
 import { usePostBulkMatchOrCategorize } from '@api/businesses/[business-id]/bank-transactions/bulk-match-or-categorize/post'
-import { useSelectedIds } from '@providers/bankTransactions/BulkSelectionStore/BulkSelectionStoreProvider'
 import { type BankTransactionCategorization, BankTransactionSelectionVariant, DEFAULT_CATEGORIZATION, useGetAllBankTransactionsCategorizations } from '@providers/categorization/BankTransactionsCategorizationStore/BankTransactionsCategorizationStoreProvider'
+import { useSelectedIds } from '@providers/global/BulkSelectionStore/BulkSelectionStoreProvider'
 import { useLayerContext } from '@providers/global/LayerContext/LayerContext'
 
 type MatchOrCategorizeTransaction = typeof MatchOrCategorizeTransactionRequestSchema.Type
